incorporate doug's suggestions:
(a) Can the numerical values of beta, phi_d, phi_m be written so that as we successively view scroll through the inversions at different beta then these numbers will appear. The reason for this, as we turn the corner on the tikhonov curve, the changes in misfit become small;changes in  misfit in the second decimal place give large changes in the model. This makes a big impact when teaching.

(b) (Optional and low priority): but we constantly advocate showing normalized misfits. Would that be straightforward to incorporate?

(c) the axis label for data doesn’t move with the plot. So it becomes a label for the model